plugin
set up plugin.
Set up the plugin.blade.
In this blade Im updating a template with JavaScript to allow users to select labels for text, and then marking the selected text with those labels. Right-clicking on the labeled text removes the label from the text and the associated entry from the Map that tracks the labels.