index.js,blade,plugin.php,index.css
index.js: we add the rightclick so user can delete the selected label of text, we parse the data from patterns +some other bugs.

blade:some changes so we parse the data from patterns like elementId.

plugin.php: in getViewAnnotations() we parse the data of patterns,  saveDatatoDatabase() (it doesnt work)

index.css: change the name of selected-label to annotations-label