Save to db, remove tooltip, add label name next to selected text
index.css: remove the tooltip.

index.js: we take the data from state so we can save the changes on db.  We add genColorFromText and we remove it from plugin.php. We fix some other bugs so user cant select words from selected text. Also in createSpan  we remove the tooltip and we add the name of label next to selected text. It takes color from the genColorCodeFromText.

plugin.php: we remove the function that add color so we did some more changes to the code.

plugin.blade.php: We add the state and the labels. Also we add the toUpperCase() method to save the name of labels with the correct way (uppercase).